+++ cloud-sql/mysql/Startup.cs
@@ -21,6 +21,10 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using MySql.Data.MySqlClient;
+using Polly;
+using System;
+using System.Data;
+using System.Data.Common;
 
 namespace CloudSql
 {
@@ -39,17 +43,7 @@ namespace CloudSql
         // http://go.microsoft.com/fwlink/?LinkID=398940
         public void ConfigureServices(IServiceCollection services)
         {
-            string projectId = Google.Api.Gax.Platform.Instance().ProjectId;
-            if (!string.IsNullOrEmpty(projectId))
-            {
-                services.AddGoogleExceptionLogging(options =>
-                {
-                    options.ProjectId = projectId;
-                    options.ServiceName = "CloudSqlSample";
-                    options.Version = "Test";
-                });
-            }
-            services.AddScoped(sp => new MySqlConnection().GetMySqlConnectionString());
+            services.AddSingleton(sp => new MySqlConnection().GetMySqlConnectionString());
             services.AddMvc(options =>
             {
                 options.Filters.Add(typeof(DbExceptionFilterAttribute));
@@ -79,6 +73,148 @@ namespace CloudSql
                     name: "default",
                     template: "{controller=Home}/{action=Index}/{id?}");
             });
+        }
+    }
+
+    static class StartupExtensions
+    {
+        public static DbConnection OpenWithRetry(this DbConnection connection) =>
+            // [START cloud_sql_mysql_dotnet_ado_backoff]
+            Policy
+                .HandleResult<DbConnection>(conn => conn.State != ConnectionState.Open)
+                .WaitAndRetry(new[]
+                {
+                    TimeSpan.FromSeconds(1),
+                    TimeSpan.FromSeconds(2),
+                    TimeSpan.FromSeconds(5)
+                }, (result, timeSpan, retryCount, context) =>
+                {
+                    // Log any warnings here.
+                    if(result.Exception != null && retryCount == 3)
+                    {
+                        throw result.Exception;
+                    }
+                })
+                .Execute(() =>
+                {
+                    // Return a new connection.
+                    try {
+                        connection.Open();
+                    }
+                    catch (MySqlException e)
+                    {
+                         Console.WriteLine(
+                             $"Error connecting to database: {e.Message}");
+                    }
+                    return connection;
+                });
+            // [END cloud_sql_mysql_dotnet_ado_backoff]
+
+        public static void InitializeDatabase(this DbConnection connection)
+        {
+            using(connection.OpenWithRetry())
+            {
+                using (var createTableCommand = connection.CreateCommand())
+                {
+                    createTableCommand.CommandText = @"
+                        CREATE TABLE IF NOT EXISTS
+                        votes(
+                            vote_id SERIAL NOT NULL,
+                            time_cast timestamp NOT NULL,
+                            candidate CHAR(6) NOT NULL,
+                            PRIMARY KEY (vote_id)
+                        )";
+                    createTableCommand.ExecuteNonQuery();
+                }
+            }
+        }
+
+        public static MySqlConnectionStringBuilder GetMySqlConnectionString(this DbConnection connection)
+        {
+            MySqlConnectionStringBuilder connectionString; 
+            if (Environment.GetEnvironmentVariable("DB_HOST") != null)
+            {
+                connectionString = NewMysqlTCPConnectionString();
+            }
+            else
+            {
+                connectionString = NewMysqlUnixSocketConnectionString();
+            }
+            // The values set here are for demonstration purposes only. You 
+            // should set these values to what works best for your application.
+            // [START cloud_sql_mysql_dotnet_ado_limit]
+            // MaximumPoolSize sets maximum number of connections allowed in the pool.
+            connectionString.MaximumPoolSize = 5;
+            // MinimumPoolSize sets the minimum number of connections in the pool.
+            connectionString.MinimumPoolSize = 0;
+            // [END cloud_sql_mysql_dotnet_ado_limit]
+            // [START cloud_sql_mysql_dotnet_ado_timeout]
+            // ConnectionTimeout sets the time to wait (in seconds) while
+            // trying to establish a connection before terminating the attempt.
+            connectionString.ConnectionTimeout = 15;
+            // [END cloud_sql_mysql_dotnet_ado_timeout]
+            // [START cloud_sql_mysql_dotnet_ado_lifetime]
+            // ConnectionLifeTime sets the lifetime of a pooled connection
+            // (in seconds) that a connection lives before it is destroyed
+            // and recreated. Connections that are returned to the pool are
+            // destroyed if it's been more than the number of seconds
+            // specified by ConnectionLifeTime since the connection was
+            // created. The default value is zero (0) which means the
+            // connection always returns to pool.
+            connectionString.ConnectionLifeTime = 1800; // 30 minutes
+            // [END cloud_sql_mysql_dotnet_ado_lifetime]
+            return connectionString;
+        }
+
+        public static MySqlConnectionStringBuilder NewMysqlTCPConnectionString()
+        {
+            // [START cloud_sql_mysql_dotnet_ado_connection_tcp]
+            // Equivalent connection string:
+            // "Uid=<DB_USER>;Pwd=<DB_PASS>;Host=<DB_HOST>;Database=<DB_NAME>;"
+            var connectionString = new MySqlConnectionStringBuilder()
+            {
+                // The Cloud SQL proxy provides encryption between the proxy and instance.
+                SslMode = MySqlSslMode.None,
+
+                // Remember - storing secrets in plain text is potentially unsafe. Consider using
+                // something like https://cloud.google.com/secret-manager/docs/overview to help keep
+                // secrets secret.
+                Server = Environment.GetEnvironmentVariable("DB_HOST"),   // e.g. '127.0.0.1'
+                // Set Host to 'cloudsql' when deploying to App Engine Flexible environment
+                UserID = Environment.GetEnvironmentVariable("DB_USER"),   // e.g. 'my-db-user'
+                Password = Environment.GetEnvironmentVariable("DB_PASS"), // e.g. 'my-db-password'
+                Database = Environment.GetEnvironmentVariable("DB_NAME"), // e.g. 'my-database'
+            };
+            connectionString.Pooling = true;
+            // Specify additional properties here.
+            return connectionString;
+            // [END cloud_sql_mysql_dotnet_ado_connection_tcp]
+        }
+
+        public static MySqlConnectionStringBuilder NewMysqlUnixSocketConnectionString()
+        {
+            // [START cloud_sql_mysql_dotnet_ado_connection_socket]
+            // Equivalent connection string:
+            // "Server=<dbSocketDir>/<INSTANCE_CONNECTION_NAME>;Uid=<DB_USER>;Pwd=<DB_PASS>;Database=<DB_NAME>;Protocol=unix"
+            String dbSocketDir = Environment.GetEnvironmentVariable("DB_SOCKET_PATH") ?? "/cloudsql";
+            String instanceConnectionName = Environment.GetEnvironmentVariable("INSTANCE_CONNECTION_NAME");
+            var connectionString = new MySqlConnectionStringBuilder()
+            {
+                // The Cloud SQL proxy provides encryption between the proxy and instance.
+                SslMode = MySqlSslMode.None,
+                // Remember - storing secrets in plain text is potentially unsafe. Consider using
+                // something like https://cloud.google.com/secret-manager/docs/overview to help keep
+                // secrets secret.
+                Server = String.Format("{0}/{1}", dbSocketDir, instanceConnectionName),
+                UserID = Environment.GetEnvironmentVariable("DB_USER"),   // e.g. 'my-db-user
+                Password = Environment.GetEnvironmentVariable("DB_PASS"), // e.g. 'my-db-password'
+                Database = Environment.GetEnvironmentVariable("DB_NAME"), // e.g. 'my-database'
+                ConnectionProtocol = MySqlConnectionProtocol.UnixSocket
+            };
+            connectionString.Pooling = true;
+            // Specify additional properties here.
+            return connectionString;
+            // [END cloud_sql_mysql_dotnet_ado_connection_socket]
         }
     }
 }
